First experiment for updating OSM with post-2006 TIGER changes
Currently for a single county in Indiana:

First, calls out TLIDs that are in TIGERrd13 but not tiger2006se
that should be examined and considered for inclusion in OSM.

Then looks for tiger2006se TLIDs that survive unchanged in OSM ways
and colors them to indicate that they could be directly replaced by
the TIGERrd13 equivalents without breaking anything else.

Requires my weirdo reformatting of 1-degree-square OSM extracts.