TLIDs ends that once shared the node might not any more.
When looking at the block edges for a replacement for a missing TLID,
look for all the places each TLID endpoint is known to have been
moved, not just one of them.

Always look for the block edge intersections, too, since that might
be the one that connects instead of some other place the endpoints
were moved to.